Introduce sql_base.py
